pyproject.toml
[project]
name = "aioamazondevices"
version = "11.0.2"
requires-python = ">=3.12"
description = "Python library to control Amazon devices"
authors = [
{ name = "Simone Chemelli", email = "simone.chemelli@gmail.com" },
]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/chemelli74/aioamazondevices"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries",
]
packages = [
    { include = "aioamazondevices", from = "src" },
]
dependencies = [
"aiohttp (>=3.12.7)",
"beautifulsoup4",
"colorlog",
"langcodes",
"orjson (>=3.10,<4)",
"python-dateutil",
]

[project.urls]
"Homepage" = "https://github.com/chemelli74/aioamazondevices"
"Bug Tracker" = "https://github.com/chemelli74/aioamazondevices/issues"
"Changelog" = "https://github.com/chemelli74/aioamazondevices/blob/main/CHANGELOG.md"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0"
pytest-cov = ">=5,<8"
types-python-dateutil = "^2.9.0.20251115"

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
version_variables = [
    "src/aioamazondevices/__init__.py:__version__",
]
build_command = "pip install poetry && poetry build"

[tool.semantic_release.changelog]
exclude_commit_patterns = [
    "chore*",
    "ci*",
]

[tool.semantic_release.changelog.environment]
keep_trailing_newline = true

[tool.semantic_release.branches.main]
match = "main"

[tool.semantic_release.branches.noop]
match = "(?!main$)"
prerelease = true

[tool.pytest.ini_options]
addopts = "-v -Wdefault --cov=aioamazondevices --cov-report=term-missing:skip-covered"
pythonpath = ["src"]

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "@overload",
    "if TYPE_CHECKING",
    "raise NotImplementedError",
    'if __name__ == "__main__":',
]

[tool.ruff.isort]
known-first-party = ["aioamazondevices", "tests"]

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
mypy_path = "src/"
no_implicit_optional = true
show_error_codes = true
warn_unreachable = true
warn_unused_ignores = true
exclude = [
    'setup.py',
]

[[tool.mypy.overrides]]
module = "tests.*"
allow_untyped_defs = true

[build-system]
requires = ["poetry-core>=2.0"]
build-backend = "poetry.core.masonry.api"
